Move tasks into its own folder
This shouldnt have been in the Ui theme folder, moved to data folder
