+++ app/src/main/java/com/example/FocusApp/screens/TaskScreen.kt
@@ -40,8 +40,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.em
-import com.example.FocusApp.ui.theme.Task
-import com.example.FocusApp.ui.theme.TaskListViewModel
+import com.example.FocusApp.data.Task
+import com.example.FocusApp.data.TaskListViewModel
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.font.FontStyle
